PureSpecial

Documentation for PureSpecial.

PureSpecial.PureSpecial

PureSpecial.Specfun

PureSpecial.ai_zeros

PureSpecial.bi_zeros

PureSpecial.itairy

PureSpecial.PureSpecial — Module

Special functions in Pure Julia

Elementary Functions

Gamma Functions

Exponential and Trigonometric Integrals

Error Functions

Airy Functions

Bessel Functions

Struve Functions

Parabolic Cylinder Functions

Hypergeometric Functions

Legendre Functions

Elliptic Integrals

Elliptic Functions

Zeta Functions

Mathieu Functions

Spheroidal Wave Functions

Miscellaneous Functions

PureSpecial.ai_zeros — Method

ai_zeros(nt)

Compute the first nt zeros of Airy functions Ai(x) and Ai'(x), a and a', and the associated values of Ai(a') and Ai'(a).

PureSpecial.bi_zeros — Method

bi_zeros(nt)

Compute the first nt zeros of Airy functions Bi(x) and Bi'(x), b and b', and the associated values of Bi(b') and Bi'(b).

PureSpecial.itairy — Method

itairy(x)

Compute the integrals of Airy functions with respect to t from 0 and x ( x ≥ 0 )