Autogen Docs

PureSpecial.PureSpecial

PureSpecial.Specfun

PureSpecial.ai_zeros

PureSpecial.bi_zeros

PureSpecial.digamma

PureSpecial.factorial2

PureSpecial.factorialk

PureSpecial.gamma

PureSpecial.gamma

PureSpecial.itairy

PureSpecial.logabsgamma

PureSpecial.loggamma

PureSpecial.polygamma

PureSpecial.trigamma

PureSpecial.PureSpecial — Module

Special functions in Pure Julia

Elementary Functions

Gamma Functions

Exponential and Trigonometric Integrals

Error Functions

Airy Functions

Bessel Functions

Struve Functions

Parabolic Cylinder Functions

Hypergeometric Functions

Legendre Functions

Elliptic Integrals

Elliptic Functions

Zeta Functions

Mathieu Functions

Spheroidal Wave Functions

Miscellaneous Functions

PureSpecial.ai_zeros — Method

ai_zeros(nt)

Compute the first nt zeros of Airy functions Ai(x) and Ai'(x), a and a', and the associated values of Ai(a') and Ai'(a).

PureSpecial.bi_zeros — Method

bi_zeros(nt)

Compute the first nt zeros of Airy functions Bi(x) and Bi'(x), b and b', and the associated values of Bi(b') and Bi'(b).

PureSpecial.digamma — Function

digamma(z)

Compute the digamma function (psi function).

PureSpecial.factorial2 — Function

factorial2(n)

Compute the double factorial function n!!.

PureSpecial.factorialk — Function

factorialk(n, k)

Compute the multifactorial function of n of order k, n(!!...!).

PureSpecial.gamma — Method

gamma(z)

Compute the gamma function for complex z.

Examples

julia> [ (i, gamma(i)) for i in 1:5 ]
5-element Vector{Tuple{Int64, Float64}}:
 (1, 1.0)
 (2, 1.0)
 (3, 2.0)
 (4, 6.0)
 (5, 24.0)

julia> gamma(0.5)^2 ≈ π
true

julia> gamma(4 + 1) == prod(1:4) == factorial(4)
true

See also: factorial, loggamma.

PureSpecial.itairy — Method

itairy(x)

Compute the integrals of Airy functions with respect to t from 0 and x ( x ≥ 0 )

PureSpecial.logabsgamma — Function

logabsgamma(z)

Compute the logarithm of the absolute value of the gamma function log(abs(gamma(z)))

PureSpecial.loggamma — Function

loggamma(z)

Compute the principal branch of the logarithm of the gamma function log(gamma(z)).

PureSpecial.polygamma — Function

polygamma(n, x)

Compute the polygamma function.

PureSpecial.trigamma — Function

trigamma(z)

Compute the trigamma function.