Gamma Functions

Factorial Function

Function	Description

Base.factorial	factorial function

factorial2	double factorial function

factorialk	multifactorial function

Base.binomial	binomial coefficient

Binomial Coefficient

Gamma Function

Function	Description

gamma	gamma function

loggamma	log gamma function

logabsgamma	log abs gamma function

digamma	digamma function

trigamma	trigamma function

polygamma	polygamma function

Incomplete Gamma Function

Pochhammer Function

Beta Function

Autodoc

PureSpecial.digamma — Function

digamma(z)

Compute the digamma function (psi function).

PureSpecial.factorial2 — Function

factorial2(n)

Compute the double factorial function n!!.

PureSpecial.factorialk — Function

factorialk(n, k)

Compute the multifactorial function of n of order k, n(!!...!).

PureSpecial.gamma — Method

gamma(z)

Compute the gamma function for complex z.

Examples

julia> [ (i, gamma(i)) for i in 1:5 ]
5-element Vector{Tuple{Int64, Float64}}:
 (1, 1.0)
 (2, 1.0)
 (3, 2.0)
 (4, 6.0)
 (5, 24.0)

julia> gamma(0.5)^2 ≈ π
true

julia> gamma(4 + 1) == prod(1:4) == factorial(4)
true

See also: factorial, loggamma.

PureSpecial.logabsgamma — Function

logabsgamma(z)

Compute the logarithm of the absolute value of the gamma function log(abs(gamma(z)))

PureSpecial.loggamma — Function

loggamma(z)

Compute the principal branch of the logarithm of the gamma function log(gamma(z)).

PureSpecial.polygamma — Function

polygamma(n, x)

Compute the polygamma function.

PureSpecial.trigamma — Function

trigamma(z)

Compute the trigamma function.