AutoGen Docs

Index

PureSpecial.Specfun.SF_EULER_GAMMA

PureSpecial.Specfun.SF_EULER_GAMMA_00

PureSpecial.Specfun.SF_EULER_GAMMA_28

PureSpecial.Specfun.SF_INF300

PureSpecial.Specfun.SF_PI

PureSpecial.Specfun._GAM0_G

PureSpecial.Specfun._LGAMA_A

PureSpecial.Specfun.airyb

PureSpecial.Specfun.airyzo

PureSpecial.Specfun.aswfa

PureSpecial.Specfun.bernoa

PureSpecial.Specfun.bernob

PureSpecial.Specfun.beta

PureSpecial.Specfun.bjndd

PureSpecial.Specfun.cchg

PureSpecial.Specfun.cerf

PureSpecial.Specfun.cerzo

PureSpecial.Specfun.cfc

PureSpecial.Specfun.cfs

PureSpecial.Specfun.cgama

PureSpecial.Specfun.chgm

PureSpecial.Specfun.chgm_kernel

PureSpecial.Specfun.chgu

PureSpecial.Specfun.chgubi

PureSpecial.Specfun.chguit

PureSpecial.Specfun.chgul

PureSpecial.Specfun.chgus

PureSpecial.Specfun.cisia

PureSpecial.Specfun.cisib

PureSpecial.Specfun.cpbdn

PureSpecial.Specfun.cpdla

PureSpecial.Specfun.cpdsa

PureSpecial.Specfun.cpsi

PureSpecial.Specfun.djn_zeros

PureSpecial.Specfun.dvla

PureSpecial.Specfun.dvsa

PureSpecial.Specfun.dyn_zeros

PureSpecial.Specfun.e1xa

PureSpecial.Specfun.e1xb

PureSpecial.Specfun.e1z

PureSpecial.Specfun.eix

PureSpecial.Specfun.eixz

PureSpecial.Specfun.envj

PureSpecial.Specfun.enxa

PureSpecial.Specfun.enxb

PureSpecial.Specfun.erf

PureSpecial.Specfun.erf

PureSpecial.Specfun.eulera

PureSpecial.Specfun.eulerb

PureSpecial.Specfun.fcs

PureSpecial.Specfun.fcszo

PureSpecial.Specfun.ffk

PureSpecial.Specfun.gaih

PureSpecial.Specfun.gam0

PureSpecial.Specfun.gamma2

PureSpecial.Specfun.incob

PureSpecial.Specfun.incog

PureSpecial.Specfun.itairy

PureSpecial.Specfun.itsh0

PureSpecial.Specfun.itsl0

PureSpecial.Specfun.itth0

PureSpecial.Specfun.jdzo

PureSpecial.Specfun.jdzo

PureSpecial.Specfun.jn_zeros

PureSpecial.Specfun.jynbh

PureSpecial.Specfun.jyndd

PureSpecial.Specfun.jyzo

PureSpecial.Specfun.klvna

PureSpecial.Specfun.klvnb

PureSpecial.Specfun.klvnzo

PureSpecial.Specfun.kmn

PureSpecial.Specfun.lgama

PureSpecial.Specfun.lpmns

PureSpecial.Specfun.lpmv

PureSpecial.Specfun.lpmv0

PureSpecial.Specfun.lqmns

PureSpecial.Specfun.msta1

PureSpecial.Specfun.msta2

PureSpecial.Specfun.pbdv

PureSpecial.Specfun.pbvv

PureSpecial.Specfun.pbwa

PureSpecial.Specfun.psi

PureSpecial.Specfun.rmn1

PureSpecial.Specfun.rmn2l

PureSpecial.Specfun.rmn2sp

PureSpecial.Specfun.rswfp

PureSpecial.Specfun.sckb

PureSpecial.Specfun.sdmn

PureSpecial.Specfun.segv

PureSpecial.Specfun.sphj

PureSpecial.Specfun.sphy

PureSpecial.Specfun.stvh0

PureSpecial.Specfun.stvh1

PureSpecial.Specfun.stvhv

PureSpecial.Specfun.stvl0

PureSpecial.Specfun.vvla

PureSpecial.Specfun.vvsa

PureSpecial.Specfun.yn_zeros

Docs

PureSpecial.Specfun — Module

Special functions in pure julia.

Translated from specfun.f

C
C       COMPUTATION OF SPECIAL FUNCTIONS
C
C          Shanjie Zhang and Jianming Jin
C
C       Copyrighted but permission granted to use code in programs.
C       Buy their book "Computation of Special Functions", 1996, John Wiley & Sons, Inc.
C

PureSpecial.Specfun.SF_EULER_GAMMA — Constant

Euler's Constant γ

Same as: Base.MathConstants.eulergamma

julia> Float64(Base.MathConstants.eulergamma)
0.5772156649015329

PureSpecial.Specfun.SF_EULER_GAMMA_00 — Constant

Euler's Constant γ

julia> Float64(Base.MathConstants.eulergamma)
0.5772156649015329

julia> 0.57721566490153_00  # imprecise
0.57721566490153

PureSpecial.Specfun.SF_EULER_GAMMA_28 — Constant

Euler's Constant γ

julia> Float64(Base.MathConstants.eulergamma)
0.5772156649015329

julia> 0.57721566490153_28  # imprecise
0.5772156649015328

PureSpecial.Specfun.SF_INF300 — Constant

1e300 used as +Inf in specfun.f

PureSpecial.Specfun.SF_PI — Constant

The constant π.

Same as: Base.MathConstants.pi

julia> Float64(Base.MathConstants.pi)
3.141592653589793

PureSpecial.Specfun._GAM0_G — Constant

Ref

DLMF 5.7.1,5.7.2

Wrench, J. W. (1968). Concerning two series for the gamma function. Mathematics of Computation, 22(103), 617-626. https://www.jstor.org/stable/2004538

TODO: The coefficients are not consistent with those in the paper

PureSpecial.Specfun._LGAMA_A — Constant

Coefficients for the series expansion

PureSpecial.Specfun.airyb — Method

airyb(x::Float64)

Compute Airy functions and their derivatives.

Input

x: Argument of Airy function

Output

(ai, bi, ad, bd)

ai: Ai(x)

bi: Bi(x)

ad: Ai'(x)

bd: Bi'(x)

PureSpecial.Specfun.airyzo — Method

airyzo(
    nt::Int, kf::Int, 
    xa::Vector{Float64}, xb::Vector{Float64},
    xc::Vector{Float64}, xd::Vector{Float64}
)

Compute the first NT zeros of Airy functions Ai(x) and Ai'(x), a and a', and the associated values of Ai(a') and Ai'(a); and the first NT zeros of Airy functions Bi(x) and Bi'(x), b and b', and the associated values of Bi(b') and Bi'(b).

Example

nt = 4;
a,b,c,d = zeros(nt),zeros(nt),zeros(nt),zeros(nt)
airyzo(nt, 1, a,b,c,d)
@show a b c d;

Input

NT –- Total number of zeros

KF –- Function code

KF=1 for Ai(x) and Ai'(x)

KF=2 for Bi(x) and Bi'(x)

Output

XA(m) –- a, the m-th zero of Ai(x) or b, the m-th zero of Bi(x)

XB(m) –- a', the m-th zero of Ai'(x) or b', the m-th zero of Bi'(x)

XC(m) –- Ai(a') or Bi(b')

XD(m) –- Ai'(a) or Bi'(b) ( m –- Serial number of zeros )

Routine called

Specfun.airyb for computing Airy functions and their derivatives

PureSpecial.Specfun.aswfa — Method

Compute the prolate and oblate spheroidal angular functions of the first kind and their derivatives

Input :

m –- Mode parameter, m = 0,1,2,...

n –- Mode parameter, n = m,m+1,...

c –- Spheroidal parameter

x –- Argument of angular function, |x| ≤ 1.0

KD –- Function code

KD= 1 for prolate;

KD=-1 for oblate

cv –- Characteristic value

Output: (s1f, s1d)

s1f –- Angular function of the first kind

s1d –- Derivative of the angular function of the first kind

Call:

Specfun.sdmn

Specfun.sckb

PureSpecial.Specfun.bernoa — Method

Compute Bernoulli number Bn, using recurrence relation. Not work for large n.

Input :

n –- Serial number, n >= 2

Output:

BN(n) –- Bn, n = 0..N

PureSpecial.Specfun.bernob — Method

Compute Bernoulli number Bn, using series expression (DLMF 24.8.1).

Input :

n –- Serial number, n >= 2

Output:

BN(n) –- Bn, n = 0..N

PureSpecial.Specfun.beta — Method

Compute beta function B(p, q).

Input :

p –- Parameter ( p > 0 )

q –- Parameter ( q > 0 )

Output:

B(p, q)

Routine called:

Specfun.gamma2 for computing Γ(x)

PureSpecial.Specfun.bjndd — Method

bjndd(x::Float64, n::Int)

Compute Bessel functions Jn(x) and their first and second derivatives ( 0 <= n <= 100 )

Input

x –- Argument of Jn(x) ( x ≥ 0 )

n –- Order of Jn(x)

Output

BJ(n+1) –- Jn(x)

DJ(n+1) –- Jn'(x)

FJ(n+1) –- Jn"(x)

PureSpecial.Specfun.cchg — Method

cchg(a::Float64, b::Float64, z::Complex{Float64})

Compute confluent hypergeometric function M(a, b, z) with real parameters a, b and a complex argument z.

Input

a –- Parameter

b –- Parameter

z –- Complex argument

Output

M(a,b,z)

Routine called

Specfun.cgama for computing complex ln[Г(x)]

Reference

DLMF C13: Confluent Hypergeometric Functions

Zhang, S.J., & Jin, J.M. (1996). Computation of Special Functions. Wiley.

PureSpecial.Specfun.cerf — Method

Compute complex Error function erf(z) & erf'(z)

Input

z –- Complex argument of erf(z)

Output: (cer, cder)

CER –- erf(z)

CDER –- erf'(z)

PureSpecial.Specfun.cerzo — Method

Evaluate the complex zeros of error function erf(z) using the modified Newton's iteration method

Input :

NT –- Total number of zeros

Output:

ZO(L) –- L-th zero of erf(z), L=1,2,...,NT

Routine called:

Specfun.cerf for computing erf(z) and erf'(z)

PureSpecial.Specfun.cfc — Method

Compute complex Fresnel integral C(z) and C'(z)

Input

z –- Argument of C(z)

Output: (zf, zd)

C(z)

C'(z)

PureSpecial.Specfun.cfs — Method

Compute complex Fresnel Integral S(z) and S'(z)

Input

z –- Argument of S(z)

Output: (zf, zd)

S(z)

S'(z)

PureSpecial.Specfun.cgama — Method

cgama(z::Complex{Float64}, kf::Int)

Compute complex gamma function ln[Г(z)] or Г(z).

Input

z –- Complex argument

kf –- Function code

kf=1 for Г(z)

kf=0 for ln[Г(z)]

Output

ln[Г(z)] or Г(z)

PureSpecial.Specfun.chgm — Method

chgm(a::Float64, b::Float64, x::Float64)

Compute confluent hypergeometric function M(a, b, x).

Input

a –- Parameter

b –- Parameter ( b != 0,-1,-2,... )

x –- Argument

Output

M(a,b,x)

Routine called

Specfun.cgama for computing complex ln[Г(x)]

PureSpecial.Specfun.chgm_kernel — Method

chgm_kernel(a::Float64, b::Float64, x::Float64)

F77 impl in scipy, without input check.

PureSpecial.Specfun.chgu — Method

chgu(a::Float64, b::Float64, x::Float64)

Compute the confluent hypergeometric function U(a, b, x).

Input:

a –- Parameter

b –- Parameter

x –- Argument ( x > 0 )

Output: (hu, md, isfer)

HU –- U(a,b,x)

MD –- Method code

ISFER –- Error flag

Routines called:

Specfun.chgus for small x ( MD=1 )

Specfun.chgul for large x ( MD=2 )

Specfun.chgubi for integer b ( MD=3 )

Specfun.chguit for numerical integration ( MD=4 )

PureSpecial.Specfun.chgubi — Method

chgubi(a::Float64, b::Float64, x::Float64)

Compute confluent hypergeometric function U(a, b, x) with integer b ( b = ±1,±2,... ).

Input:

a –- Parameter

b –- Parameter

x –- Argument

Output: (hu, id)

HU –- U(a,b,x)

ID –- Estimated number of significant digits

Routines called:

Specfun.gamma2 for computing gamma function Г(x)

Specfun.psi for computing psi function

PureSpecial.Specfun.chguit — Method

chguit(a::Float64, b::Float64, x::Float64)

Compute hypergeometric function U(a, b, x) by using Gaussian-Legendre integration (n = 60).

Input:

a –- Parameter ( a > 0 )

b –- Parameter

x –- Argument ( x > 0 )

Output: (hu, id)

HU –- U(a,b,z)

ID –- Estimated number of significant digits

Routine called:

Specfun.gamma2 for computing Г(x)

PureSpecial.Specfun.chgul — Method

chgul(a::Float64, b::Float64, x::Float64)

Compute the confluent hypergeometric function U(a, b, x) for large argument x.

Input:

a –- Parameter

b –- Parameter

x –- Argument, x > 0

Output: (hu, id)

HU –- U(a,b,x)

ID –- Estimated number of significant digits

PureSpecial.Specfun.chgus — Method

chgus(a::Float64, b::Float64, x::Float64)

Compute confluent hypergeometric function U(a, b, x) for small argument x.

Input:

a –- Parameter

b –- Parameter ( b ≠ 0, -1, -2, ...)

x –- Argument

Output: (hu, id)

HU –- U(a,b,x)

ID –- Estimated number of significant digits

Routine called:

Specfun.gamma2 for computing gamma function

PureSpecial.Specfun.cisia — Method

cisia(x::Float64)

Compute cosine and sine integrals Ci(x) and Si(x).

Parameters:

x: Argument of Ci(x) and Si(x), x ≥ 0

Returns: (ci, si)

Ci(x)

Si(x)

PureSpecial.Specfun.cisib — Method

cisib(x::Float64)

Compute cosine and sine integrals Ci(x) and Si(x), using polynomial and rational approximations:

x <= 1, |eps(x)| <= 1e-7

1 < x <= Inf, |eps_ci(x)| <= 5e-7, |eps_si(x)| <= 3e-7

Parameters:

x: Argument of Ci(x) and Si(x), x ≥ 0

Returns: (ci, si)

Ci(x)

Si(x)

PureSpecial.Specfun.cpbdn — Method

Compute the parabolic cylinder functions Dn(z) and Dn'(z) for a complex argument.

Input:

z –- Complex argument of Dn(z)

n –- Order of Dn(z) ( n=0, ±1, ±2, … )

Output: (cpb, cpd)

CPB(|n|) –- Dn(z)

CPD(|n|) –- Dn'(z)

Routine called:

Specfun.cpdsa for computing Dn(z) for a small |z|

Specfun.cpdla for computing Dn(z) for a large |z|

PureSpecial.Specfun.cpdla — Method

Compute complex parabolic cylinder function Dn(z) for large argument.

Input:

z –- Complex argument of Dn(z), Re(z) > 0, |z| >> n, z -> +Inf

n –- Order of Dn(z) (n = 0, ±1, ±2,…)

Output:

cdn –- Dn(z)

PureSpecial.Specfun.cpdsa — Method

Compute complex parabolic cylinder function Dn(z) for small argument.

Input:

z –- Complex argument of Dn(z), Re(z) <= 7

n –- Order of Dn(z) (n = 0, -1, -2,…)

Output:

cdn –- Dn(z)

Routine called:

Specfun.gaih for computing Г(x), x = n/2 (n = 1, 2, ...)

PureSpecial.Specfun.cpsi — Method

cpsi(z::ComplexF64)

Compute complex psi function Ψ(z) (Digamma Function).

Input :

z

Output:

psi(z)

PureSpecial.Specfun.djn_zeros — Method

djn_zeros(rj1::Vector{Float64}, n::Int64, nt::Int64)

Compute zeros of integer-order Bessel function Jn(x) and its derivatives Jn'(x)

Input

n –- Order of Bessel functions (n >= 0)

NT –- Number of zeros (roots)

Output

RJ0(L) –- L-th zero of Jn(x), L=1,2,...,NT

RJ1(L) –- L-th zero of Jn'(x), L=1,2,...,NT

Routine called

Specfun.jyndd for computing Jn(x), and its first and second derivatives

PureSpecial.Specfun.dvla — Method

Compute parabolic cylinder functions Dv(x) for large argument

Input

x –- Argument

va –- Order

Output

Dv(x)

Routines called:

Specfun.vvla for computing Vv(x) for large |x|

Specfun.gamma2 for computing Г(x)

PureSpecial.Specfun.dvsa — Method

Compute parabolic cylinder function Dv(x) for small argument

Input

x –- Argument

va –- Order

Output

Dv(x)

Routine called

Specfun.gamma2 for computing Г(x)

PureSpecial.Specfun.dyn_zeros — Method

dyn_zeros(ry1::Vector{Float64}, n::Int64, nt::Int64)

Compute zeros of integer-order Bessel function Yn(x) and its derivatives Yn'(x)

Input

n –- Order of Bessel functions (n >= 0)

NT –- Number of zeros (roots)

Output

RY0(L) –- L-th zero of Yn(x), L=1,2,...,NT

RY1(L) –- L-th zero of Yn'(x), L=1,2,...,NT

Routine called

Specfun.jyndd for computing Yn(x), and its first and second derivatives

PureSpecial.Specfun.e1xa — Method

e1xa(x::Float64)

Compute exponential integral E1(x), using rational approximation:

0 < x <= 1: |eps(x)| <= 2e10-7

1 < x: |eps(x)| <= 2e10-8

Input:

x: Argument of E1(x), ( x > 0 )

Output:

E1(x)

PureSpecial.Specfun.e1xb — Method

e1xb(x::Float64)

Compute exponential integral E1(x).

Input

x: Argument of E1(x), ( x > 0 )

Output

E1(x)

PureSpecial.Specfun.e1z — Method

e1z(z::Complex{Float64})

Compute complex exponential integral E1(z).

Input

z: Argument of E1(z)

Output

E1(z)

PureSpecial.Specfun.eix — Method

eix(x::Float64)

Compute exponential integral Ei(x).

Input

x: Argument of Ei(x)

Output

Ei(x)

PureSpecial.Specfun.eixz — Method

eixz(z::Complex{Float64})

Compute complex exponential integral Ei(z).

Input

x: Complex argument of Ei(z)

Output

Ei(z)

PureSpecial.Specfun.envj — Method

envj(n, x)

Helper function used in msta1, msta2

PureSpecial.Specfun.enxa — Method

enxa(n::Int, x::Float64)

Compute exponential integral En(x), using forward recurrence.

0 < x <= 20: |eps(x)| <= 1e-8

Parameters:

n : Order of En(x), n >= 1

x : Argument of En(x), ( 0 < x <= 20 )

Returns:

[ En(x) for n in 0:n ] :: Vector{Float64}

Routine called:

Specfun.e1xb for computing E1(x)

PureSpecial.Specfun.enxb — Method

enxb(n::Int, x::Float64)

Compute exponential integral En(x).

Parameters:

n: Order of En(x), n >= 1

x: Argument of En(x), ( x >= 0 )

Returns:

[ En(x) for n in 0:n ] :: Vector{Float64}

PureSpecial.Specfun.erf — Method

Compute error function erf(z) for a complex argument (z=x+iy)

Input:

z –- Complex argument

Output:

erf(z)

PureSpecial.Specfun.erf — Method

Compute error function erf(x)

Input:

x –- Argument of erf(x)

Output:

erf(x)

PureSpecial.Specfun.eulera — Method

Compute Euler number En, using recurrence relation. Not work for large n.

Input :

n –- Serial number, n >= 2

Output:

EN(n) –- En, n = 0..N

PureSpecial.Specfun.eulerb — Method

Compute Euler number En, using series expression (DLMF 24.8.4).

Input :

n –- Serial number, n >= 2

Output:

EN(n) –- En, n = 0..N

PureSpecial.Specfun.fcs — Method

Compute Fresnel integrals C(x) and S(x)

Input

x –- Argument of C(x) and S(x)

Output (c, s)

C(x)

S(x)

PureSpecial.Specfun.fcszo — Method

Compute the complex zeros of Fresnel integral C(z) or S(z) using modified Newton's iteration method

Input

KF –- Function code

KF=1 for C(z)

KF=2 for S(z)

NT –- Total number of zeros

Output

ZO(L) –- L-th zero of C(z) or S(z)

Routines called:

Specfun.cfc for computing Fresnel integral C(z)

Specfun.cfs for computing Fresnel integral S(z)

PureSpecial.Specfun.ffk — Method

Compute modified Fresnel integrals F±(x) and K±(x)

Input

x –- Argument of F±(x) and K±(x)

KS –- Sign code KS=0 for calculating F+(x) and K+(x) KS=1 for calculating F(x) and K(x)

Output

FR –- Re[F±(x)]

FI –- Im[F±(x)]

FM –- |F±(x)|

FA –- Arg[F±(x)], Degs.

GR –- Re[K±(x)]

GI –- Im[K±(x)]

GM –- |K±(x)|

GA –- Arg[K±(x)], Degs.

PureSpecial.Specfun.gaih — Method

gaih(x::Float64)

Compute gamma function Г(x)

Input x –- Argument of Г(x), x = n/2, n=1,2,3,…

Output ga –- Г(x)

PureSpecial.Specfun.gam0 — Method

gam0(x)

Compute gamma function Г(x) for small x

Input x –- Argument of Г(x) ( |x| ≤ 1 )

Output ga –- Г(x)

PureSpecial.Specfun.gamma2 — Method

gamma2(x::Float64)

Compute gamma function Г(x).

Input

x –- Argument of Г(x) ( x is not equal to 0,-1,-2,… )

Output

Г(x)

PureSpecial.Specfun.incob — Method

Compute incomplete beta function Ix(a,b).

Input :

a –- Parameter

b –- Parameter

x –- Argument ( 0 ≤ x ≤ 1 )

Output:

BIX –- Ix(a,b)

Routine called:

Specfun.beta for computing beta function B(p,q)

PureSpecial.Specfun.incog — Method

Compute incomplete gamma function γ(a,x), Γ(a,x) and P(a,x).

Input :

a –- Parameter ( a ≤ 170 )

x –- Argument

Output: (gin, gim, gip, isfer):

gin –- γ(a,x), incomplete gamma function

gim –- Γ(a,x), incomplete gamma function

gip –- P(a,x), normalized incomplete gamma function

isfer –- Error flag, isfer > 0 when error occurs

Routine called:

Specfun.gamma2

PureSpecial.Specfun.itairy — Method

itairy(x::Float64)

Compute the integrals of Airy functions with respect to t from 0 and x ( x ≥ 0 )

Input

x –- Upper limit of the integral

Output

apt –- Integration of Ai(t) from 0 and x

bpt –- Integration of Bi(t) from 0 and x

ant –- Integration of Ai(-t) from 0 and x

bnt –- Integration of Bi(-t) from 0 and x

PureSpecial.Specfun.itsh0 — Method

Evaluate the integral of Struve function H0(t) with respect to t from 0 and x

Input :

x –- Upper limit ( x ≥ 0 )

Output:

TH0 –- Integration of H0(t) from 0 and x

PureSpecial.Specfun.itsl0 — Method

Evaluate the integral of modified Struve function L0(t) with respect to t from 0 to x

Input :

x –- Upper limit ( x ≥ 0 )

Output:

TL0 –- Integration of L0(t) from 0 to x

PureSpecial.Specfun.itth0 — Method

Evaluate the integral H0(t)/t with respect to t from x to infinity

Input :

x –- Lower limit ( x ≥ 0 )

Output:

TTH –- Integration of H0(t)/t from x to infinity

PureSpecial.Specfun.jdzo — Method

jdzo(
    nt::Int,
    zo::Vector{Float64}, n::Vector{Int}, m::Vector{Int}, p::Vector{Int}
)

Compute the zeros of Bessel functions Jn(x) and Jn'(x), and arrange them in the order of their magnitudes.

Input

NT –- Number of total zeros ( NT ≤ 1200 )

Output

ZO(L) –- Value of the L-th zero of Jn(x) and Jn'(x)

N(L) –- n, order of Jn(x) or Jn'(x) associated with the L-th zero

M(L) –- m, serial number of the zeros of Jn(x) or Jn'(x) associated with the L-th zero ( L is the serial number of all the zeros of Jn(x) and Jn'(x) )

P(L) –- 0 (TM) or 1 (TE), a code for designating the zeros of Jn(x) or Jn'(x). In the waveguide applications, the zeros of Jn(x) correspond to TM modes and those of Jn'(x) correspond to TE modes

Routine called

Specfun.bjndd for computing Jn(x), Jn'(x) and Jn''(x)

PureSpecial.Specfun.jdzo — Method

jdzo(nt::Int64)

PureSpecial.Specfun.jn_zeros — Method

jn_zeros(rj0::Vector{Float64}, n::Int64, nt::Int64)

Compute zeros of integer-order Bessel functions Jn.

Input

n –- Order of Bessel functions (n >= 0)

NT –- Number of zeros (roots)

Output

RJ0(L) –- L-th zero of Jn(x), L=1,2,...,NT

Routine called

Specfun.jyndd for computing Jn(x), and its first and second derivatives

PureSpecial.Specfun.jynbh — Method

jynbh(
    x::Float64, n::Int, nmin::Int,
    bj::Vector{Float64}, by::Vector{Float64}
)

Compute Bessel functions Jn(x), Yn(x)

Input

x –- Argument of Jn(x) and Yn(x) ( x ≥ 0 )

n –- Highest order of Jn(x) and Yn(x) computed ( n ≥ 0 )

nmin – Lowest order computed ( nmin ≥ 0 )

Output

BJ(n-NMIN) –- Jn(x) ; if indexing starts at 0

BY(n-NMIN) –- Yn(x) ; if indexing starts at 0

Return:

NM –- Highest order computed

Routines called

Specfun.msta1 and Specfun.msta2 to calculate the startingpoint for backward recurrence

PureSpecial.Specfun.jyndd — Method

jyndd(x::Float64, n::Int)

compute bessel functions jn(x) and yn(x), and their first and second derivatives.

input

x –- argument of jn(x) and yn(x) ( x > 0 )

n –- order of jn(x) and yn(x)

output

bjn –- jn(x)

djn –- jn'(x)

fjn –- jn"(x)

byn –- yn(x)

dyn –- yn'(x)

fyn –- yn"(x)

routines called

Specfun.jynbh to compute jn and yn

PureSpecial.Specfun.jyzo — Method

jyzo(
    n::Int64, nt::Int64,
    rj0::Vector{Float64}, rj1::Vector{Float64},
    ry0::Vector{Float64}, ry1::Vector{Float64}
)

Compute the zeros of Bessel functions Jn(x), Yn(x), and their derivatives

Input

n –- Order of Bessel functions (n >= 0)

NT –- Number of zeros (roots)

Output

RJ0(L) –- L-th zero of Jn(x), L=1,2,...,NT

RJ1(L) –- L-th zero of Jn'(x), L=1,2,...,NT

RY0(L) –- L-th zero of Yn(x), L=1,2,...,NT

RY1(L) –- L-th zero of Yn'(x), L=1,2,...,NT

Routine called

Specfun.jyndd for computing Jn(x), Yn(x), and their first and second derivatives

PureSpecial.Specfun.klvna — Method

klvna(x::Float64)

Compute Kelvin functions ber x, bei x, ker x and kei x, and their derivatives ( x > 0 )

Input

x –- Argument of Kelvin functions

Output

ber –- ber x

bei –- bei x

ger –- ker x

gei –- kei x

der –- ber'x

dei –- bei'x

her –- ker'x

hei –- kei'x

PureSpecial.Specfun.klvnb — Function

Compute Kelvin functions and their derivatives using polynomial approximation.

Use Specfun.klvna instead.

PureSpecial.Specfun.klvnzo — Method

klvnzo(nt::Int, kd::Int)

Compute the zeros of Kelvin functions.

Input

NT –- Total number of zeros

KD –- Function code

KD=1 to 8 for ber x, bei x, ker x, kei x, ber'x, bei'x, ker'x and kei'x, respectively.

Output

ZO(M) –- the M-th zero of Kelvin function for code KD.

Routine called

Specfun.klvna for computing Kelvin functions and their derivatives.

PureSpecial.Specfun.kmn — Method

Compute the expansion coefficients of the prolate and oblate spheroidal functions and joining factors.

Input:

df, in, read-only

Output:

dn

Return: (ck1, ck2)

PureSpecial.Specfun.lgama — Method

lgama(kf::Int, x::Float64)

Compute gamma function ln[Γ(x)] or Γ(x).

Input:

x –- Argument of Γ(x) ( x > 0 )

kf –- Function code

kf=0 for ln[Γ(x)]

kf=1 for Γ(x)

Output:

ln[Γ(x)] or Γ(x)

PureSpecial.Specfun.lpmns — Method

Compute associated Legendre functions Pmn(x) and Pmn'(x) for a given order.

Input:

m –- Order of Pmn(x), m = 0,1,2,...,n

n –- Degree of Pmn(x), n = 0,1,2,...,N

x –- Argument of Pmn(x)

Output:

Pmn(x)

Pmn'(x)

PureSpecial.Specfun.lpmv — Method

Compute the associated Legendre function Pmv(x) with an integer order and an arbitrary degree v, using recursion for large degrees.

Input:

v –- Degree of Pmv(x)

m –- Order of Pmv(x)

x –- Argument of Pm(x), -1 ≤ x ≤ 1

Output:

Pmv(x)

Routine called:

lpmv0

gamma2

PureSpecial.Specfun.lpmv0 — Method

Compute the associated Legendre function Pmv(x) with an integer order and an arbitrary nonnegative degree v.

Input:

v –- Degree of Pmv(x), v >= 0

m –- Order of Pmv(x)

x –- Argument of Pm(x), -1 ≤ x ≤ 1

Output:

Pmv(x)

Routine called:

Specfun.psi for computing Psi function

PureSpecial.Specfun.lqmns — Method

Compute associated Legendre functions Qmn(x) and Qmn'(x) for a given order.

Input:

m –- Order of Qmn(x), m = 0,1,2,...

n –- Degree of Qmn(x), n = 1,2,...

x –- Argument of Qmn(x)

Output:

Qmn(x)

Qmn'(x)

PureSpecial.Specfun.msta1 — Method

msta1(x::Float64, mp::Int)

Determine the starting point for backward recurrence such that the magnitude of Jn(x) at that point is about 10^(-MP)

Input

x –- Argument of Jn(x)

MP –- Value of magnitude

Output

Starting point

PureSpecial.Specfun.msta2 — Method

msta2(x::Float64, n::Int64, mp::Int64)

Determine the starting point for backward recurrence such that all Jn(x) has MP significant digits

Input

x –- Argument of Jn(x)

n –- Order of Jn(x)

MP –- Significant digit

Output

Starting point

PureSpecial.Specfun.pbdv — Method

Compute parabolic cylinder functions Dv(x) and their derivatives

Input

x –- Argument of Dv(x)

v –- Order of Dv(x)

Output

DV(na) –- Dn+v0(x)

DP(na) –- Dn+v0'(x) ( na = |n|, v0 = v-n, |v0| < 1, n = 0,±1,±2,… )

(pdf, pdd)

PDF –- Dv(x)

PDD –- Dv'(x)

Routines called:

Specfun.dvsa for computing Dv(x) for small |x|

Specfun.dvla for computing Dv(x) for large |x|

PureSpecial.Specfun.pbvv — Method

Compute parabolic cylinder functions Vv(x) and their derivatives

Input

x –- Argument of Vv(x)

v –- Order of Vv(x)

Output

VV(na) –- Vv(x)

VP(na) –- Vv'(x) ( na = |n|, v = n+v0, |v0| < 1 n = 0,±1,±2,… )

(pvf, pvd)

PVF –- Vv(x)

PVD –- Vv'(x)

Routines called:

Specfun.vvsa for computing Vv(x) for small |x|

Specfun.vvla for computing Vv(x) for large |x|

PureSpecial.Specfun.pbwa — Method

Compute parabolic cylinder functions W(a,±x) and their derivatives

Input

a –- Parameter ( 0 ≤ |a| ≤ 5 )

x –- Argument of W(a,±x) ( 0 ≤ |x| ≤ 5 )

Output

(w1f, w1d, w2f, w2d)

W1F –- W(a,x)

W1D –- W'(a,x)

W2F –- W(a,-x)

W2D –- W'(a,-x)

Routine called:

Specfun.cgama for computing complex gamma function

PureSpecial.Specfun.psi — Method

Compute Psi function Ψ(x) (Digamma Function).

Input:

x –- Argument of psi(x)

Output:

psi(x)

PureSpecial.Specfun.rmn1 — Method

Compute prolate and oblate spheroidal radial functions of the first kind for given m, n, c, and x

Routines called:

Specfun.sckb

Specfun.sphj

PureSpecial.Specfun.rmn2l — Method

Compute prolate and oblate spheroidal radial functions of the second kind for given m, n, c and a large cx

Routine called:

Specfun.sphy for computing the spherical Bessel

functions of the second kind

PureSpecial.Specfun.rmn2sp — Method

Compute prolate spheroidal radial function of the second kind with a small argument

Routines called:

Specfun.kmn for computing expansion coefficients and joining factors

Specfun.lpmns for computing the associated Legendre functions of the first kind

Specfun.lqmns for computing the associated Legendre functions of the second kind

PureSpecial.Specfun.rswfp — Method

Compute prolate spheriodal radial functions of the first and second kinds, and their derivatives

Input: m –- Mode parameter, m = 0,1,2,... n –- Mode parameter, n = m,m+1,m+2,... c –- Spheroidal parameter x –- Argument of radial function ( x > 1.0 ) cv –- Characteristic value KF –- Function code KF=1 for the first kind KF=2 for the second kind KF=3 for both the first and second kinds

Output: R1F –- Radial function of the first kind R1D –- Derivative of the radial function of the first kind R2F –- Radial function of the second kind R2D –- Derivative of the radial function of the second kind

Routines called:

Specfun.sdmn

Specfun.rmn1

Specfun.rmn2l

rmn2sp

PureSpecial.Specfun.sckb — Method

Compute the expansion coefficients of the prolate and oblate spheroidal functions

Input:

m –- Mode parameter

`n –- Mode parameter

c –- Spheroidal parameter

DF(k) –- Expansion coefficients dk

Output:

CK(k) –- Expansion coefficients ck; CK(1), CK(2), ... correspond to c0, c2, ...

PureSpecial.Specfun.sdmn — Method

Compute the expansion coefficients of the prolate and oblate spheroidal functions, dk

Input:

m –- Mode parameter

n –- Mode parameter

c –- Spheroidal parameter

cv –- Characteristic value

KD –- Function code

KD= 1 for prolate;

KD=-1 for oblate

Output:

DF(k) –- Expansion coefficients dk; DF(1), DF(2), ... correspond to d0, d2, ... for even n-m and d1, d3, ... for odd n-m

PureSpecial.Specfun.segv — Method

Compute the characteristic values of spheroidal wave functions

Input:

m –- Mode parameter

n –- Mode parameter

c –- Spheroidal parameter

KD –- Function code

KD=1 for Prolate;

KD=-1 for Oblate

Output:

CV –- Characteristic value for given m, n and c

EG(L) –- Characteristic value for mode m and n' ( L = n' - m + 1 )

PureSpecial.Specfun.sphj — Method

Compute spherical Bessel functions jn(x) and their derivatives

Input: x –- Argument of jn(x) n –- Order of jn(x) ( n = 0,1,… )

Output: sj –- Array of jn(x) values dj –- Array of jn'(x) values nm –- Highest order computed

Routines called:

Specfun.msta1

Specfun.msta2

PureSpecial.Specfun.sphy — Method

Compute spherical Bessel functions yn(x) and their derivatives

Input :

x –- Argument of yn(x) ( x ≥ 0 )

n –- Order of yn(x) ( n = 0,1,… )

Output:

SY(n) –- yn(x)

DY(n) –- yn'(x)

NM –- Highest order computed

PureSpecial.Specfun.stvh0 — Method

Compute Struve Function H0(x)

Input:

x –- Argument of H0(x) ( x ≥ 0 )

Output:

H0(x)

PureSpecial.Specfun.stvh1 — Method

Compute Struve Function H1(x)

Input:

x –- Argument of H1(x) ( x ≥ 0 )

Output:

H1(x)

PureSpecial.Specfun.stvhv — Method

Compute Struve Functions Hv(x) with arbitrary order v ( -8.0 ≤ v ≤ 12.5 )

Input :

v –- Order of Hv(x)

x –- Argument of Hv(x) ( x ≥ 0 )

Output :

Hv(x)

Required:

Specfun.gamma2 function to compute the Gamma function

PureSpecial.Specfun.stvl0 — Method

Compute Modified Struve Function L0(x)

Input:

x –- Argument of L0(x) ( x ≥ 0 )

Output:

SL0 –- L0(x)

PureSpecial.Specfun.vvla — Method

Compute parabolic cylinder function Vv(x) for large argument

Input

x –- Argument

va –- Order

Output

Vv(x)

Routines called:

Specfun.dvla for computing Dv(x) for large |x|

Specfun.gamma2 for computing Г(x)

PureSpecial.Specfun.vvsa — Method

Compute parabolic cylinder function Vv(x) for small argument

Input

x –- Argument

va –- Order

Output

Vv(x)

Routine called

Specfun.gamma2 for computing Г(x)

PureSpecial.Specfun.yn_zeros — Method

yn_zeros(ry0::Vector{Float64}, n::Int64, nt::Int64)

Compute zeros of integer-order Bessel function Yn(x).

Input

n –- Order of Bessel functions (n >= 0)

NT –- Number of zeros (roots)

Output

RY0(L) –- L-th zero of Yn(x), L=1,2,...,NT

Routine called

Specfun.jyndd for computing Yn(x), and its first and second derivatives